webinar service

+++ src/main/java/xcode/ilmugiziku/domain/response/invoice/InvoiceResponse.java
@@ -3,9 +3,11 @@ package xcode.ilmugiziku.domain.response.invoice;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class InvoiceResponse {
     private String packageName;
-    private int totalAmount;
+    private BigDecimal totalAmount;
 }
